requirements.txt
asgiref==3.3.1
Django==3.1.4
django-rest-registration==0.5.6
djangorestframework==3.12.2
icalendar==4.0.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
